WG-54781 waapi-client-python users can't guaranty order of publish events
Added an indirection on publishers in the form of a callback executor
that the client can optionally specify.

Three are implemented:

* PerCallbackThreadExecutor: Spawns a new thread on each publish
* SequentialThreadExecutor: Schedules on a thread for sequential
  execution
* AsyncioLoopExecutor: Schedules on the global asyncio loop

The default has switched from PerCallbackThreadExecutor to
SequentialThreadExecution so that order is guaranteed, at the cost of no
event handler being run concurrently.
AsyncioLoopExecutor does not support using the WAAPI client insinde a
handler as calls are synchronous. Adding async versions of call and
subscribe would enable this, but there is currently no intention to
support both paradigm in the same client implementation.

Change-Id: Ia5fb16b2ecb8d5828985fe00fdc9e877c0bd69a5